fix(Data): CoreDataBus 时间戳转为北京时间并修复溢出风险
1. TimeHelper.ToUtcDateTime 乘法改为 double 中间量，避免 long 溢出

2. LowFrequencyPublisher UTC 时间加 8 小时转为北京时间，格式改为 +08:00

3. PersistenceService CSV 持久化时间戳同步转为北京时间

+++ ConsoleApp1/Tools/TimeHelper.cs
@@ -17,7 +17,7 @@ public static class TimeHelper
         long frequency)
     {
         long elapsedTicks = sampleTick - referenceTick;
-        long elapsed100ns = elapsedTicks * 10_000_000L / frequency;
+        long elapsed100ns = (long)((double)elapsedTicks * 10_000_000L / frequency);
         return new DateTime(referenceUtcTicks + elapsed100ns, DateTimeKind.Utc);
     }
 }
